+++ src/main/java/delfos/group/results/groupevaluationmeasures/diversity/ils/IntraListSimilarity.java
@@ -245,12 +245,12 @@ public class IntraListSimilarity extends GroupEvaluationMeasure {
                     continue;
                 }
 
-                List<Float> item1Features = svdModel.getItemFeatures(item1.getId()).stream().map(value -> ((Number) value).floatValue()).collect(Collectors.toList());
-                List<Float> item2Features = svdModel.getItemFeatures(item2.getId()).stream().map(value -> ((Number) value).floatValue()).collect(Collectors.toList());
+                List<Double> item1Features = svdModel.getItemFeatures(item1.getId()).stream().map(value -> ((Number) value).doubleValue()).collect(Collectors.toList());
+                List<Double> item2Features = svdModel.getItemFeatures(item2.getId()).stream().map(value -> ((Number) value).doubleValue()).collect(Collectors.toList());
 
-                float similarity = cosineCoefficient.similarity(item1Features, item2Features);
+                double similarity = cosineCoefficient.similarity(item1Features, item2Features);
 
-                if (Float.isFinite(similarity)) {
+                if (Double.isFinite(similarity)) {
                     sumOfSimilarities += similarity;
                 } else {
                     sumOfSimilarities += 1;
